Follow File Name Conventions
File naming conventions now state that all PHP files (except templates) should
be lower-case and use underscores to separate words in a filename.

There is another pull request in place that modifies IslandoraBookReader.js
It may be best to wait till its pulled and then merge with this before pulling.