Page flip correctly appears moving from left to right rather than rigt to left.
This is a bug that only effects small books.

A list of prefetched images is maintained by the book reader, when the book
wraps in a small book the first two images are still in this list, and they
maintain their old inline styles. These previous settings were messing with
the display of the page flip making the page appear to be flipping in the
opposite direction.